feat: ManyToOneバインディング修正による運動記録の保存機能完了
原因：ManyToOneリレーションシップにおけるTypeMismatchエラーにより、フォームが正常に保存されず画面に戻ってしまう問題を修正しました。

- Exerciseエンティティに一時的な`exerciseTypeId`（@Transient）フィールドを追加し、フォームからのIDを正しく受け取れるように変更。
- HTMLのselectタグを新しいフィールド名(`name="exerciseTypeId"`)に対応するよう修正。
- Controller内で、一時的なIDを使って完全な`ExerciseType`エンティティを手動でロードし、`Exercise`オブジェクトにセットしてからDBに保存するように修正。

+++ src/main/java/com/springboot/healthmanage/mapper/ExerciseTypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ExerciseTypeRepository extends JpaRepository<ExerciseType, Long> {
+public interface ExerciseTypeRepository extends JpaRepository<ExerciseType, Integer> {
 
     /**
      * 获取所有运动类型，并按名称排序
