+++ src/main/java/com/springboot/healthmanage/service/ExerciseTypeService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ExerciseTypeService {
 
-    ExerciseType findExerciseTypeById(Long exerciseTypeId);
+    ExerciseType findExerciseTypeById(Integer exerciseTypeId);
 
     List<ExerciseType> findAllExerciseTypes();
     List<String> findAllExerciseTypeNames();
